chore: commit message can be multiline and you can't have multiline links
testing multiline comments

because multiline comments are likely

also likely are really long comments. Because, well, who doesn't like a loooong comment? LooooooOOOOOong comments tell a story, probably not a good story, but a story. You might be changing a bunch of stuff and need lots of words to explain what you're doing, or maybe you just know a lot of words and want the world to know who awesome you are? (fyi; you're not awesome)